+++ src/main/java/com/qa/main/controllers/AnimalController.java
@@ -4,6 +4,7 @@ import java.util.List;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -17,7 +18,7 @@ import com.qa.main.entities.Animal;
 import com.qa.main.services.AnimalService;
 
 @RestController
-
+@CrossOrigin
 @RequestMapping("/animal")
 public class AnimalController {
