+++ src/main/resources/static/Axios/index.html
@@ -1,26 +1,119 @@
 <!DOCTYPE html>
 <html lang="en">
   <head>
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Document</title>
+    <title>Caniformia CyberPark</title>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.6.1/dist/css/bootstrap.min.css">
+    <script src="https://cdn.jsdelivr.net/npm/jquery@3.6.0/dist/jquery.slim.min.js"></script>
+    <script src="https://cdn.jsdelivr.net/npm/popper.js@1.16.1/dist/umd/popper.min.js"></script>
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@4.6.1/dist/js/bootstrap.bundle.min.js"></script>
+    <style>
+      body {
+        background-image: url("./imgs/fadedfox.jpg");
+        background-color: #76e998;
+        background-size: cover;
+      }
+    </style>
   </head>
-  <body onLoad = "getAll()">
-    <!-- CSS only -->
-    <link
-      href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.1/dist/css/bootstrap.min.css"
-      rel="stylesheet"
-      integrity="sha384-iYQeCzEYFbKjA/T2uDLTpkwGzCiq6soy8tYaI1GyVh/UjpbCx/TYkiZhlZB6+fzT"
-      crossorigin="anonymous"
-    />
+  <header>
+    <nav class="navbar navbar-expand-lg navbar-light bg-light">
+      <a class="navbar-brand" href="#">Navbar</a>
+      <button
+        class="navbar-toggler"
+        type="button"
+        data-toggle="collapse"
+        data-target="#navbarSupportedContent"
+        aria-controls="navbarSupportedContent"
+        aria-expanded="false"
+        aria-label="Toggle navigation">
+        <span class="navbar-toggler-icon"></span>
+      </button>
+      <div class="collapse navbar-collapse" id="navbarSupportedContent">
+        <ul class="navbar-nav mr-auto">
+          <li class="nav-item active">
+            <a class="nav-link" href="./index.html"
+              >Home <span class="sr-only">(current)</span></a>
+          </li>
+          <li class="nav-item">
+            <h5><a class="nav-link" href="https://www.wwf.org.uk/">WWF</a></h5>
+          </li>
+          <li class="nav-item dropdown">
+            <a class="nav-link dropdown-toggle" href="#" id="navbarDropdown" role="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
+              Site Quick Nav
+            </a>
+            <div class="dropdown-menu" aria-labelledby="navbarDropdown">
+              <a class="dropdown-item" href="./index.html">Home</a>
+              <a class="dropdown-item" href="./indexF.html">Infosheet</a>
+              <div class="dropdown-divider"></div>
+              <a class="dropdown-item" href="./account.html">Sign up</a>
+            </div>
+          </li>
+          <li class="nav-item">
+            <a class="nav-link disabled" href="#"></a>
+          </li>
+        </ul>
+        <form class="form-inline my-2 my-lg-0">
+          <input class="form-control mr-sm-2" type="search" placeholder="Search" aria-label="Search"/>
+          <button class="btn btn-outline-success my-2 my-sm-0" type="submit">Search</button>
+        </form>
+      </div>
+    </nav>
+  </header>
+  <body onLoad="getAll()">
+  <div id="formDiv">
+     <h3>Animal Entry</div></h3>
+      <div class="card" style="width: 18rem;">         
+          <ul class="list-group list-group-flush">
+            <li class="list-group-item"><form action="/action_page.php" class="was-validated">
+              <div class="mb-3 mt-3">
+                  <label for="cname" class="form-label">Common Name:</label>
+                  <input type="text" class="form-control" id="cname" placeholder="Enter name" name="cname" required>
+                  <div class="valid-feedback">Valid.</div>
+                  <div class="invalid-feedback">Please fill out this field.</div>
+              </div></li>
+            <li class="list-group-item">
+              <label for="exampleDataList" class="form-label">Genus</label>
+              <input class="form-control" list="datalistOptions" id="exampleDataList" placeholder="Choose from dropdown...">
+              <datalist id="datalistOptions">
+                <option value="Vulpes vulpes">
+                <option value="Vulpes lagopus"> 
+                <option value="Vulpes zerda">
+                <option value="Canis lupus">
+                <option value="Canis lupus arctos">
+                <option value="Nyctereutes procyonoides">
+              </datalist></li>
+            <li class="list-group-item"><form action="/action_page.php" class="was-validated">
+              <div class="mb-3">
+                  <label for="sname" class="form-label">Size:</label>
+                  <input type="number" class="form-control" id="sname" placeholder="Enter size" name="sname" required>
+                  <div class="valid-feedback">Valid.</div>
+                  <div class="invalid-feedback">Please fill out this field.</div>
+              </div></li>
+            <li class="list-group-item"><select class="form-select" size="3" aria-label="size 3 select example">
+              <option selected>Open this select menu</option>
+              <option value="1">Canidae</option>
+              <option value="2">Two</option>
+              <option value="3">Three</option>
+            </select>
+              </div></li>
+          </ul>
+          <div class="card-footer">
+          <button type="button" id="createBtn" class="btn btn-outline-success">Create</button>
+          <button type="button" id="updateBtn" class="btn btn-outline-primary">Update</button>
+          <button type="button" id="deleteBtn" class="btn btn-outline-danger">Delete</button>
+      </div>
+</div>
+
     <!-- JavaScript Bundle with Popper -->
     <script
       src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.1/dist/js/bootstrap.bundle.min.js"
       integrity="sha384-u1OknCvxWvY5kfmNBILK2hRnQC3Pr17a+RTT6rIHI7NnikvbZlHgTPOOmMi466C8"
       crossorigin="anonymous"
     ></script>
+    <!-- Axios JS -->
+    <script src="https://cdn.jsdelivr.net/npm/axios/dist/axios.min.js"></script>
 
-    <script src="./index.js"></script>
+    <script src="./js/main.js"></script>
   </body>
 </html>
