+++ src/main/resources/static/Axios/indexF.html
@@ -1,0 +1,110 @@
+<html lang="en">
+  <head>
+    <title>Caniformia CyberPark</title>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.6.1/dist/css/bootstrap.min.css">
+    <script src="https://cdn.jsdelivr.net/npm/jquery@3.6.0/dist/jquery.slim.min.js"></script>
+    <script src="https://cdn.jsdelivr.net/npm/popper.js@1.16.1/dist/umd/popper.min.js"></script>
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@4.6.1/dist/js/bootstrap.bundle.min.js"></script>
+    <style> 
+      body {
+        background-image: url("./imgs/fadedfox.jpg");
+        background-color: #76e998;
+        background-size: cover;
+      }
+      </style>
+  </head>
+  <header>
+    <nav class="navbar navbar-expand-lg navbar-light bg-light">
+        <a class="navbar-brand" href="#">Navbar</a>
+        <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
+          <span class="navbar-toggler-icon"></span>
+        </button>
+      
+        <div class="collapse navbar-collapse" id="navbarSupportedContent">
+          <ul class="navbar-nav mr-auto">
+            <li class="nav-item active">
+              <a class="nav-link" href="./index.html">Home <span class="sr-only"></span></a>
+            </li>
+            <li class="nav-item">
+              <h5><a class="nav-link" href="https://www.wwf.org.uk/">WWF</a></h5>
+            </li>
+            <li class="nav-item dropdown">
+              <a class="nav-link dropdown-toggle" href="#" id="navbarDropdown" role="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
+                Site Quick Nav</a>
+              <div class="dropdown-menu" aria-labelledby="navbarDropdown">
+                <a class="dropdown-item" href="./index.html">Home</a>
+                <a class="dropdown-item" href="./indexF.html">Infosheets</a>
+                <div class="dropdown-divider"></div>
+                <a class="dropdown-item" href="./account.html">Sign up</a>
+              </div>
+            </li>
+            <li class="nav-item">
+              <a class="nav-link disabled" href="#"></a>
+            </li>
+          </ul>
+          <form class="form-inline my-2 my-lg-0">
+            <input class="form-control mr-sm-2" type="search" placeholder="Search" aria-label="Search">
+            <button class="btn btn-outline-success my-2 my-sm-0" type="submit">Search</button>
+          </form>
+        </div>
+      </nav>
+  </header>
+  <body>
+    <h1>Foxes - Info sheets</h1>
+    <div class="container mx-auto mt-4">
+      <div class="row">
+        <div class="col-md-4">
+          <div class="card" style="width: 19rem">
+            <img src="./imgs/ZTkt4I5.jpg" class="card-img-top" alt="..." />
+            <div class="card-body">
+              <h5 class="card-title">Red Fox</h5>
+              <i><h6 class="card-subtitle mb-2 text-muted">Vulpes vulpes</h6></i>
+              <p class="card-text">
+                The red fox is the largest of the true foxes and one of the most widely distributed members of the order Carnivora, being present across the entire Northern Hemisphere including most of North America, Europe and Asia, plus parts of North Africa. It is listed as least concern by the IUCN. Its range has increased alongside human expansion, having been introduced to Australia, where it is considered harmful to native mammals and bird populations.
+              </p>
+              <a href="https://en.wikipedia.org/wiki/Red_fox" class="btn btn-primary mr-2"
+                ><i class="fas fa-link"></i> Visit Site</a>
+              <a href="https://www.wildlifetrusts.org/get-involved" class="btn btn-primary"
+                ><i class="fab fa-github"></i> Get involved</a>
+          </div>
+          </div>
+        </div>
+        <div class="col-md-4">
+          <div class="card" style="width: 21rem">
+            <img src="./imgs/arctic.jpg" class="card-img-top" alt="..." />
+            <div class="card-body">
+              <h5 class="card-title">Arctic Fox</h5>
+              <i><h6 class="card-subtitle mb-2 text-muted">Vulpes lagopus</h6></i>
+              <p class="card-text">
+                The Arctic fox, also known as the white fox, polar fox, or snow fox, is a small fox native to the Arctic regions of the Northern Hemisphere and common throughout the Arctic tundra biome. It is well adapted to living in cold environments, and is best known for its thick, warm fur that is also used as camouflage. It has a large and very fluffy tail. In the wild, most individuals do not live past their first year but some exceptional ones survive up to 11 years. Its body length ranges from 46 to 68 cm (18 to 27 in), with a generally rounded body shape to minimize the escape of body heat.
+              </p>
+              <a href="https://en.wikipedia.org/wiki/Arctic_fox" class="btn btn-primary mr-2"
+                ><i class="fas fa-link"></i> Visit Site</a>
+                <a href="https://www.wildlifetrusts.org/about-us" class="btn btn-primary"
+                ><i class="fab fa-github"></i> About the Trust</a>
+            </div>
+          </div>
+        </div>
+        <div class="col-md-4">
+          <div class="card" style="width: 22rem">
+            <img src="./imgs/fennec.jpg" class="card-img-top" alt="..." />
+            <div class="card-body">
+              <h5 class="card-title">Fennec Fox</h5>
+              <i><h6 class="card-subtitle mb-2 text-muted">Vulpes zerda</h6></i>
+              <p class="card-text">
+                The fennec fox is a small crepuscular fox native to the deserts of North Africa, ranging from Western Sahara to the Sinai Peninsula. Its most distinctive feature is its unusually large ears, which serve to dissipate heat and listen for underground prey. The fennec is the smallest species of fox. Its coat, ears, and kidney functions have adapted to the desert environment with high temperatures and little water. It mainly eats insects, small mammals, and birds. The fennec has a life span of up to 14 years in captivity and about 10 years in the wild. Its main predators are the Verreaux's eagle-owl, jackals, and other large mammals.
+              </p>
+              <a href="https://en.wikipedia.org/wiki/Fennec_fox" class="btn btn-primary mr-2"
+                ><i class="fas fa-link"></i> Visit Site</a>
+                <a href="https://www.wildlifetrusts.org/visit" class="btn btn-primary"
+                ><i class="fab fa-github"></i> Visit nature</a>
+            </div>
+          </div>
+        </div>
+        </div>
+      </div>
+    </div>
+  </body>
+</html>
